Now loading Hero animation from TMX data
That allowed for removing all the animation loading code that
wasn't based on TMX, as well as the assets it was loading.

Currently, the LPC-based hero animation I'm using is too big for
the map tiles I have, but I'm still considering wither to get
bigger (And nicer) tiles, or to make the character smaller.

Some more things to explore:
- We have a hero item on the map in order to make it include the
  animation. Perhaps make it the hero spawn point.
- Better tile cropping as part of the asset pipeline, as we don't
  really need the full 64x64 pixels we have ATM for the hero
  animation.
- Start breaking apart `main.py` to individual files.